fix np float

+++ patterns/swirl.py
@@ -11,7 +11,7 @@ class Swirl(Pattern):
 
     def cache_positions(self, leds):
         self._led_deltas = np.array(list(led.coord.get_delta("global") for led in leds), dtype=np.float32)
-        self._led_thetas = np.array(list(led.coord.get("global", "spherical").theta for led in leds), dtype=np.float)
+        self._led_thetas = np.array(list(led.coord.get("global", "spherical").theta for led in leds), dtype=np.float32)
         self._theta_offsets = self._led_thetas * 0.5
         self._theta_offsets += self._led_deltas
